Information regarding theta......
theta has calculated for each half of 2*pi cycle(as the curve is changing for every half cycle) and divided with half of the points on the periodic cycle. Ceil() is a pre-defined function which converts decimal to the nearest large integer as per our requirement.  So, I have used it.  We can use even floor() function which converts decimal to the nearest small integer.